# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /models:
    get:
      operationId: listModels
      tags: [models]
      summary: List user's financial models
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinancialModelSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createModel
      tags: [models]
      summary: Create a new financial model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFinancialModelData"
      responses:
        "201":
          description: Model created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialModel"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /models/{id}:
    get:
      operationId: getModel
      tags: [models]
      summary: Get a financial model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Model details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialModel"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateModel
      tags: [models]
      summary: Update a financial model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFinancialModelData"
      responses:
        "200":
          description: Model updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialModel"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteModel
      tags: [models]
      summary: Delete a financial model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Model deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /models/{id}/duplicate:
    post:
      operationId: duplicateModel
      tags: [models]
      summary: Duplicate a financial model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Model duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialModel"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /models/{id}/archive:
    post:
      operationId: archiveModel
      tags: [models]
      summary: Archive a financial model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Model archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialModel"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /models/{id}/consultant:
    get:
      operationId: getConsultantAnalysis
      tags: [models]
      summary: Run CFO consultant analysis on a financial model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultant analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultantOutput"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /models/{id}/export:
    get:
      operationId: exportModel
      tags: [models]
      summary: Export financial model as Excel workbook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Excel file download
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /models/{id}/export/pro-forma-pdf:
    get:
      operationId: exportProFormaPdf
      tags: [models]
      summary: Export financial model as Pro Forma PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /models/{id}/export/loan-readiness-pdf:
    get:
      operationId: exportLoanReadinessPdf
      tags: [models]
      summary: Export loan readiness report as PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    ModelStatus:
      type: string
      enum: [draft, complete, archived]
    SchoolStage:
      type: string
      enum: [new_school, operating_school]
    FundingProfile:
      type: string
      enum: [tuition_based, charter_public_funded, hybrid_mixed]
...
    ExpenseRow:
      type: object
      properties:
        id:
          type: string
        category:
          $ref: "#/components/schemas/ExpenseCategory"
        lineItem:
          type: string
        enabled:
          type: boolean
        driverType:
          $ref: "#/components/schemas/ExpenseDriverType"
        amounts:
          type: array
          items:
            type: number
        note:
          type: string
      required:
        - id
        - category
        - lineItem
        - enabled
        - driverType
        - amounts
...
    PriorYearSnapshot:
      type: object
      properties:
        endingEnrollment:
          type: integer
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        endingCash:
          type: number
    CreateFinancialModelData:
      type: object
      properties:
        name:
          type: string
        currentStep:
          type: integer
        schoolStage:
          $ref: "#/components/schemas/SchoolStage"
        fundingProfile:
          $ref: "#/components/schemas/FundingProfile"
        data:
          $ref: "#/components/schemas/ModelFormData"
      required:
        - name
        - data
    UpdateFinancialModelData:
      type: object
      properties:
        name:
          type: string
        currentStep:
          type: integer
        status:
          $ref: "#/components/schemas/ModelStatus"
        schoolStage:
          $ref: "#/components/schemas/SchoolStage"
        fundingProfile:
          $ref: "#/components/schemas/FundingProfile"
        data:
          $ref: "#/components/schemas/ModelFormData"
      required:
        - data
    ModelFormData:
      type: object
      properties:
        schoolProfile:
          $ref: "#/components/schemas/SchoolProfile"
        enrollment:
          $ref: "#/components/schemas/Enrollment"
        revenue:
          $ref: "#/components/schemas/Revenue"
        staffing:
          $ref: "#/components/schemas/Staffing"
        facilities:
          $ref: "#/components/schemas/Facilities"
        revenueRows:
          type: array
          items:
            $ref: "#/components/schemas/RevenueRow"
        staffingRows:
          type: array
          items:
            $ref: "#/components/schemas/StaffingRow"
        expenseRows:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseRow"
        capitalAndDebtRows:
          type: array
          items:
            $ref: "#/components/schemas/CapitalDebtRow"
        priorYearSnapshot:
          $ref: "#/components/schemas/PriorYearSnapshot"
    SchoolProfile:
      type: object
      properties:
        schoolName:
          type: string
        state:
          type: string
        schoolType:
          type: string
          enum: [charter_school, homeschool_coop, learning_pod, microschool, private_school, tutoring_center, other]
        schoolTypeOther:
          type: string
        entityType:
          type: string
          enum: [sole_practitioner, llc_single, llc_partnership, c_corp, s_corp, nonprofit_501c3]
        ein:
          type: string
        schoolStage:
          $ref: "#/components/schemas/SchoolStage"
        fundingProfile:
          $ref: "#/components/schemas/FundingProfile"
        openingYear:
          type: integer
        currentStudents:
          type: integer
        maxCapacity:
          type: integer
        fiscalYearStartMonth:
          type: integer
        isPartialFirstYear:
          type: boolean
        year1OperatingMonths:
          type: integer
    Enrollment:
      type: object
      properties:
        year1:
          type: integer
        year2:
          type: integer
        year3:
          type: integer
        year4:
          type: integer
        year5:
          type: integer
    Revenue:
      type: object
      properties:
        tuitionPerStudent:
          type: number
        esaRevenuePerStudent:
          type: number
        publicFundingPerStudent:
          type: number
        otherRevenuePerStudent:
          type: number
        scholarshipRate:
          type: number
        annualFundraising:
          type: number
        annualDonations:
          type: number
        foundationGrants:
          type: number
        capitalGifts:
          type: number
        annualTuitionIncrease:
          type: number
    Staffing:
      type: object
      properties:
        studentsPerTeacher:
          type: number
        teacherSalary:
          type: number
        adminStaffCount:
          type: integer
        adminSalary:
          type: number
        founderSalary:
          type: number
        benefitsRate:
          type: number
...
    RevenueComposition:
      type: object
      properties:
        tuitionPct:
          type: number
        publicPct:
          type: number
        philanthropyPct:
          type: number
      required:
        - tuitionPct
        - publicPct
        - philanthropyPct
    CostComposition:
      type: object
      properties:
        staffingPctOfRevenue:
          type: number
        facilityPctOfRevenue:
          type: number
        totalOpexPctOfRevenue:
          type: number
      required:
        - staffingPctOfRevenue
        - facilityPctOfRevenue
        - totalOpexPctOfRevenue
    CumulativeYear:
      type: object
      properties:
        year:
          type: integer
        cumulativeNetIncome:
          type: number
        reserveMonths:
          type: number
      required:
        - year
        - cumulativeNetIncome
        - reserveMonths
    StressScenario:
      type: object
      properties:
        scenario:
          type: string
        y1NetIncome:
          type: number
        y5NetIncome:
          type: number
        breakEvenYear:
          type: integer
          nullable: true
      required:
        - scenario
        - y1NetIncome
        - y5NetIncome
        - breakEvenYear
    ConsultantOutput:
      type: object
      properties:
        executiveSummary:
          type: string
        biggestStrength:
          type: string
        biggestRisk:
          type: string
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/ConsultantRecommendation"
        lenderReadiness:
          type: string
          enum: [Strong, Needs Work, Not Yet Ready]
        lenderReadinessExplanation:
          type: string
        keyMetrics:
          type: array
          items:
            $ref: "#/components/schemas/ConsultantKeyMetric"
        revenueComposition:
          type: array
          items:
            $ref: "#/components/schemas/RevenueComposition"
        costComposition:
          type: array
          items:
            $ref: "#/components/schemas/CostComposition"
        cumulativeFinancials:
          type: array
          items:
            $ref: "#/components/schemas/CumulativeYear"
        stressTests:
          type: array
          items:
            $ref: "#/components/schemas/StressScenario"
        enrollmentGuidance:
          type: array
          items:
            type: string
        generatedAt:
          type: string
          format: date-time
      required:
        - executiveSummary
        - biggestStrength
        - biggestRisk
        - recommendations
        - lenderReadiness
        - lenderReadinessExplanation
        - keyMetrics
        - revenueComposition
        - costComposition
        - cumulativeFinancials
        - stressTests
        - enrollmentGuidance
        - generatedAt
    ConsultantRecommendation:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
      required:
        - title
        - description
        - priority
    ConsultantKeyMetric:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
        status:
          type: string
          enum: [good, warning, danger]
        interpretation:
          type: string
      required:
        - name
        - value
        - status
        - interpretation
    FinancialModelSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          $ref: "#/components/schemas/ModelStatus"
        currentStep:
          type: integer
        schoolStage:
          $ref: "#/components/schemas/SchoolStage"
        fundingProfile:
          $ref: "#/components/schemas/FundingProfile"
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - updatedAt
        - createdAt
    FinancialModel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          $ref: "#/components/schemas/ModelStatus"
        currentStep:
          type: integer
        schoolStage:
          $ref: "#/components/schemas/SchoolStage"
        fundingProfile:
          $ref: "#/components/schemas/FundingProfile"
        data:
          $ref: "#/components/schemas/ModelFormData"
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - data
        - updatedAt
        - createdAt